requirements.txt
transformers
datasets
outlines[transformers]
streamlit
streamlit_code_editor
python-dotenv
openai
